5 June 2015
The new file is nice, but it is still hard to distinguish a) which batch
line it corresponds with b) which playlist index it corresponds with. YT
doesn't require vids of a playlist to have a different name, it has its
url to distinguish them. I've been downloading some university vids and
been unable to continue because I deleted the batch file before renaming
the vid with that index. So I was downloading the same video a couple
times just in case.

Still some frission.