7 June 2015 ff_run
To this point, I always had a two run policy with ff_run. Once to see if
there was more than one hyphen in a filename. It doesn't handle more than
one with the current algorithm. And, if I'm changing them by hand, why
improve it :p ? Eventually, I'll use the fsm class in there to make the
parsing more extensible.

Today (then), I embedded the hyphen check and wait on input if some need
to be changed. That way, I needn't run and reedit. I'll get to the other
improvement evenually.