clean ryd
Filenames starting with the run number is way better to check off which
failed and need to go in the next run, or see if the video
moved/banned/etc. Unfortunately, now they all needed to have the run
number cut so they sort for real. This script renames them and moves the
playlist index to the back.

I also use the subprocess module. This is the next direction for run_yd.
This way I can harvest the run's exit code and forgo using it in the
filename or cleaning afterward.