feat: 190815_2319 go can use json config
The go implementation expects and utilizes a json file, conforming
to the example json schemas, to supply the static arguments for a
given flag. It currently expects this file to have the name
yt_config.json and to exist in the current directory. This version
offers no fallback for cases where the file doesn't exist.